Number of elements

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -16,6 +16,6 @@ size_t listint_len(const listint_t *h)
 		h = h->next;
 		i++;
 	}
-	printf("-> %d elements\n", i);
+	printf("-> %d elements\n", (int) i);
 	return (i);
 }
